#571: the restore suites share one harness
`testRestoreClamp.js` (#558), `testRestoreBackDoor.js` (#559),
`testRestoreResolutionChanged.js` (#560) and `testRestoreNormalization.js`
(#561) each opened with the same forty lines: the `hicDataset` mock factory,
the deferred `HICBrowser` import it forces, a stated 800x800 viewport, the
three stubs, and an embed. Copy four is past the rule of three.

`test/utils/restoreFixture.js` now holds the viewport -- with the ADR-0009
fact 5 reason attached to it -- and `restoreFixture()`, which registers the
stubs and hands back `embed` and `restore`. Each suite keeps what is its own:
its zooms, its saved states, its claims, and its header essay.

The `vi.mock` call itself cannot move: vitest hoists it to the top of the file
it is written in and registers it against that file's module graph. What moved
is the shape its factory returns, `datasetModule()`, which lives beside the
dataset in `test/utils/restoreDataset.js` rather than in the fixture -- a
factory that reached into the fixture would pull `hicBrowser.js` back in behind
it and deadlock the load it is part of. `testRestoreGolden.js` returns the same
shape, so it uses it too.
